requirements.txt
python-dotenv
azure-core
azure-search-documents==11.4.0b9
openai
pandas
openpyxl